+++ apps/sim/app/(auth)/login/page.tsx
@@ -1,16 +1,9 @@
-import { redirect } from 'next/navigation'
-import { isAuthDisabled } from '@/lib/core/config/feature-flags'
 import { getOAuthProviderStatus } from '@/app/(auth)/components/oauth-provider-checker'
 import LoginForm from '@/app/(auth)/login/login-form'
 
-// Force dynamic rendering to avoid prerender errors with search params
 export const dynamic = 'force-dynamic'
 
 export default async function LoginPage() {
-  if (isAuthDisabled) {
-    redirect('/workspace')
-  }
-
   const { githubAvailable, googleAvailable, isProduction } = await getOAuthProviderStatus()
 
   return (
