+++ apps/sim/app/(auth)/signup/page.tsx
@@ -1,15 +1,10 @@
-import { redirect } from 'next/navigation'
-import { isAuthDisabled, isRegistrationDisabled } from '@/lib/core/config/feature-flags'
+import { isRegistrationDisabled } from '@/lib/core/config/feature-flags'
 import { getOAuthProviderStatus } from '@/app/(auth)/components/oauth-provider-checker'
 import SignupForm from '@/app/(auth)/signup/signup-form'
 
 export const dynamic = 'force-dynamic'
 
 export default async function SignupPage() {
-  if (isAuthDisabled) {
-    redirect('/workspace')
-  }
-
   if (isRegistrationDisabled) {
     return <div>Registration is disabled, please contact your admin.</div>
   }
