+++ apps/sim/lib/auth/index.ts
@@ -1,8 +1,4 @@
 export type { AnonymousSession } from './anonymous'
-export {
-  ANONYMOUS_USER,
-  ANONYMOUS_USER_ID,
-  createAnonymousSession,
-  ensureAnonymousUserExists,
-} from './anonymous'
+export { createAnonymousSession, ensureAnonymousUserExists } from './anonymous'
 export { auth, getSession, signIn, signUp } from './auth'
+export { ANONYMOUS_USER, ANONYMOUS_USER_ID } from './constants'
